labor systems

+++ contracts/src/components/labor.cairo
@@ -3,7 +3,7 @@ struct Labor {
     #[key]
     entity_id: u128,
     #[key]
-    resource_type: u128,
+    resource_type: u8,
     balance: u64,
     last_harvest: u64,
     multiplier: u64,
